docs: api markdown code linker
