+++ package.json
@@ -100,7 +100,7 @@
     "@typescript-eslint/parser": "^2.27.0",
     "autoprefixer": "^9.7.6",
     "babel-eslint": "^10.1.0",
-    "babel-jest": "^25.2.6",
+    "babel-jest": "^26.0.1",
     "babel-plugin-transform-define": "^2.0.0",
     "babel-plugin-transform-object-assign": "^6.22.0",
     "eslint": "^6.8.0",
@@ -137,7 +137,7 @@
     "gatsby-transformer-sharp": "2.2.23",
     "hast-util-select": "^4.0.0",
     "husky": "^4.2.3",
-    "jest": "^25.2.7",
+    "jest": "^26.0.1",
     "lint-staged": "^10.1.2",
     "postcss-color-mod-function": "^3.0.3",
     "postcss-custom-media": "^7.0.8",
@@ -149,9 +149,11 @@
     "rehype-stringify": "^7.0.0",
     "remark": "^12.0.0",
     "remark-html": "^11.0.1",
+    "remark-parse": "^8.0.2",
     "stylelint": "^13.3.0",
     "stylelint-config-standard": "^20.0.0",
-    "typescript": "^3.8.3"
+    "typescript": "^3.8.3",
+    "unist-util-remove-position": "^2.0.1"
   },
   "husky": {
     "hooks": {
